Number Theoretic Transforms

Number Theoretic Transform

NumberTheoreticTransforms.ntt — Function

ntt(g, q, x::Array{T,1}) -> Array{T,1}

The Number Theoretic Transform transforms data in a similar fashion to DFT, but instead complex roots of unity it uses integer roots with all operation defined in a finite field (modulo an integer number).

ntt function implements Number Theoretic Transform directly from the formula, so it is flexible about choosing transformation params but lacks performance.

$\bar{x}_k = \sum_{n=1}^N{x_n g^{(n-1)(k-1)} } \mod q$

There is also a few constraints on choosing parameteres and input length to ensure that inverse exists and equals to the original input. (TODO: list constraints)

The arguments of ntt function are

g: transform power base, must have inversion modulo q

q: defines modulo arithmetic (all operations are done 'mod q')

x: input data, its elements must be smaller than q

NumberTheoreticTransforms.intt — Function

intt(g, q, y::Array{T,1}) -> Array{T,1}

Inverse Number Theoretic Transform implementation directly from the formula.

$x_k = N^{-1} \sum_{n=1}^N{\bar{x}_n g^{-(n-1)(k-1)} } \mod q$

The same input parameters constraints as for ntt function must be applied

Mersenne Transforms

NumberTheoreticTransforms.mnt — Function

mnt(p, x)

NumberTheoreticTransforms.imnt — Function

imnt(p, y)

Fermat Number Transform

Missing docstring.

Missing docstring for NumberTheoreticTransforms.fnt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NumberTheoreticTransforms.ifnt. Check Documenter's build log for details.